lint: no same name file and folder
