+++ zk_toolbox/crates/zk_inception/src/main.rs
@@ -42,10 +42,10 @@ struct Inception {
 pub enum InceptionSubcommands {
     /// Ecosystem related commands
     #[command(subcommand, alias = "e")]
-    Ecosystem(EcosystemCommands),
+    Ecosystem(Box<EcosystemCommands>),
     /// Chain related commands
     #[command(subcommand, alias = "c")]
-    Chain(ChainCommands),
+    Chain(Box<ChainCommands>),
     /// Prover related commands
     #[command(subcommand, alias = "p")]
     Prover(ProverCommands),
@@ -119,8 +119,8 @@ async fn main() -> anyhow::Result<()> {
 
 async fn run_subcommand(inception_args: Inception, shell: &Shell) -> anyhow::Result<()> {
     match inception_args.command {
-        InceptionSubcommands::Ecosystem(args) => commands::ecosystem::run(shell, args).await?,
-        InceptionSubcommands::Chain(args) => commands::chain::run(shell, args).await?,
+        InceptionSubcommands::Ecosystem(args) => commands::ecosystem::run(shell, *args).await?,
+        InceptionSubcommands::Chain(args) => commands::chain::run(shell, *args).await?,
         InceptionSubcommands::Prover(args) => commands::prover::run(shell, args).await?,
         InceptionSubcommands::Server(args) => commands::server::run(shell, args)?,
         InceptionSubcommands::Containers(args) => commands::containers::run(shell, args)?,
